requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.6.3
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
cssselect==1.0.3
decorator==4.3.0
Django==1.10
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-pure-pagination==0.3.0
djangorestframework==3.8.2
future==0.16.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipython==6.5.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.12.1
Jinja2==2.10
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.0
numpy==1.15.2
parsel==1.5.0
parso==0.3.1
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
Pygments==2.2.0
PyHamcrest==1.9.0
PyMySQL==0.9.2
pyOpenSSL==18.0.0
pytz==2018.5
requests==2.19.1
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.23
w3lib==1.19.0
wcwidth==0.1.7
zope.interface==4.5.0
